shuffling of methods
